test_api/requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.19.1
setuptools==71.1.0
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
wheel==0.43.0
